#51 App.tsxの内容をTopコンポーネントに移動
